backend/requirements.txt
fastapi>=0.118.0
uvicorn[standard]>=0.32.0
pydantic>=2.11.0
anthropic>=0.39.0
python-dotenv>=1.0.1
